src/benchmarks_ts/requirements.txt
numpy==1.26.4
pandas==2.2.0
rdkit==2023.9.5
pika==1.3.2
python-dotenv==1.0.1
rich==13.7.1
pyyaml==6.0.1
tqdm>=4.65
useful-rdkit-utils>=0.2.7
pillow>=9.5
sqlitedict